+++ azzure/src/App.js
@@ -6,48 +6,36 @@ import Sidebar from './sidebar';
 import Mainbar from './mainbar';
 import { useState } from 'react';
 
-function App(props) {
-    const [user, setUser] = useState(null);
-    const [loading, setLoading] = useState(true);
 
-    async function getUser() {
-        return new Promise(async (resolve, reject) => {
-            let username = null;
-            if (document.cookie) {
-                console.log(document.cookie);
-                // get cookie
-                const sessionCookie = document.cookie.split('; ').find(row => row.startsWith('sessionCookie='));
-                const cookieValue = sessionCookie.split('=')[1];
-            
-                console.log(cookieValue);
-                if (cookieValue) {
-                    // request to /api/user/cookie/:cookie
-                    const response = await fetch(`http://localhost:8001/api/user/cookie/${cookieValue}`);
-                    if (response.status === 200) {
-                        username = response.text();
-                    }
-
-                    console.log(username);
+async function getUser() {
+    return new Promise(async (resolve, reject) => {
+        let username = null;
+        if (document.cookie) {
+            console.log(document.cookie);
+            // get cookie
+            const sessionCookie = document.cookie.split('; ').find(row => row.startsWith('sessionCookie='));
+            const cookieValue = sessionCookie.split('=')[1];
+        
+            console.log(cookieValue);
+            if (cookieValue) {
+                // request to /api/user/cookie/:cookie
+                const response = await fetch(`http://localhost:8001/api/user/cookie/${cookieValue}`);
+                if (response.status === 200) {
+                    username = response.text();
                 }
+
+                console.log(username);
             }
-            resolve(username);
-        })
-    }
+        }
+        resolve(username);
+    })
+}
 
-    const username = getUser().then(username => {
-        setUser(username);
-        setLoading(false);
-    });
-
-    if (loading) {
-        return (
-            <div className="lds-dual-ring"></div>
-        )
-    }
-
+function App(props) {
+    console.log("App logged in as: ", props.user);
     return (
         <>
-            <NavbarBasicExample user={user} />
+            <NavbarBasicExample user={props.user} />
             <Container fluid>
                 <Row>
                     <Mainbar {...props} />
@@ -57,4 +45,4 @@ function App(props) {
     );
 }
 
-export default App;
+export { App, getUser };
