+++ azzure/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
+import {App, getUser} from './App';
 import reportWebVitals from './reportWebVitals';
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import "https://kit.fontawesome.com/68fc273f7b.js";
@@ -9,20 +9,35 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import NotFound from './notfound';
 import Login from './login';
 import Register from './register';
+import { useState } from 'react';
 
 function Router() {
+    const [user, setUser] = useState(null);
+    const [loading, setLoading] = useState(true);
+
+    const username = getUser().then(username => {
+        setUser(username);
+        setLoading(false);
+    });
+
+    if (loading) {
+        return (
+            <div className="lds-dual-ring"></div>
+        )
+    }
+
     return (
         <React.StrictMode>
             <BrowserRouter>
                 <Routes>
-                    <Route index path="/home" element={<App  page='home' />}/>
+                    <Route index path="/home" element={<App  page='home' user={user} />}/>
                     <Route path="/create" element={<App page='create' />}>
-                        <Route path=":id" element={<App page='create' id={this} /> }/>
+                        <Route path=":id" element={<App page='create' id={this} user={user} /> }/>
                     </Route>
-                    <Route path="/manage" element={<App page='manage' />}/>
-                    <Route path="/login" element={<Login />} />
-                    <Route path="register" element={<Register />}/>
-                    <Route path="/" element={<App page='home' />}/>
+                    <Route path="/manage" element={<App page='manage' user={user} />}/>
+                    <Route path="/login" element={<Login user={user} />} />
+                    <Route path="register" element={<Register user={user}/>}/>
+                    <Route path="/" element={<App page='home' user={user}/>}/>
                     <Route path="/*" element={<NotFound />}/>
                 </Routes>
             </BrowserRouter>
